pyproject.toml
[project]
name = "tarobot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiogram>=3.20.0.post0",
    "aiopg>=1.4.0",
    "dotenv>=0.9.9",
    "loguru>=0.7.3",
    "peewee-async>=1.1.0",
    "pewee>=0.1.0.dev9",
    "pyyaml>=6.0.2",
    "requests>=2.32.4",
]
